+++ packages/thread/src/checkUpdates.ts
@@ -85,8 +85,13 @@ export async function checkUpdatesAndLog(matterbridge: SharedMatterbridge, log: 
   const { getGitHubUpdate } = await import('@matterbridge/utils/github-version');
 
   const branch = matterbridge.matterbridgeVersion.includes('-dev-') ? 'dev' : 'main';
+  const params =
+    `?id=${encodeURIComponent(matterbridge.uuid)}` +
+    `&v=${encodeURIComponent(matterbridge.matterbridgeVersion)}` +
+    `&node=${encodeURIComponent(process.version)}` +
+    `&os=${encodeURIComponent(process.platform)}`;
   try {
-    const updateJson = await getGitHubUpdate(branch, 'update.json', 5_000);
+    const updateJson = await getGitHubUpdate(branch, 'update.json' + params, 5_000);
     log.debug(`GitHub ${branch} update status: ${debugStringify(updateJson)}.`);
     if (
       isValidString(branch === 'main' ? updateJson.latestMessage : updateJson.devMessage, 1) &&